backend/requirements.txt
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.10.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.8.3
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
distlib==0.4.0
distro==1.9.0
einops==0.8.1
fastapi==0.116.1
filelock==3.18.0
filetype==1.2.0
fsspec==2025.7.0
ftfy==6.3.1
google-auth==2.40.3
google-genai==1.29.0
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
identify==2.6.12
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
markdown2==2.5.4
markdownify==1.1.0
marker-pdf==1.8.3
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
nodeenv==1.9.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==1.99.4
opencv-python-headless==4.12.0.88
packaging==25.0
pandas==2.3.1
pdftext==0.6.3
pillow==10.4.0
platformdirs==4.3.8
pre_commit==4.2.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.4
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.47.2
surya-ocr==0.15.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.53.3
triton==3.4.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.33.1
wcwidth==0.2.13
websockets==15.0.1
